refactor: split conversation viewer controllers
