+++ src/aiSessions/conversation/viewerTarget.ts
@@ -1,0 +1,13 @@
+'use strict';
+
+import type { AiSessionProviderId } from '../../models';
+
+export interface ConversationViewerTarget {
+    projectId: string;
+    provider: AiSessionProviderId;
+    sessionId: string;
+    interactionId: string;
+    expectedRevision: string;
+    displayName: string;
+    duplicateDisplayName: boolean;
+}
